requirements.txt
streamlit
pandas
numpy
wordcloud
plotly
matplotlib
nltk
pillow